+++ ParameterSet/src/types.cc
@@ -737,8 +737,8 @@ bool
 {
   std::vector<std::string> tokens = edm::pset::tokenize(from, ":");
   assert(tokens.size() == 2);
-  unsigned long run = strtoul(tokens[0].c_str(), 0, 0);
-  unsigned long event = strtoul(tokens[1].c_str(), 0, 0);
+  unsigned int run = strtoul(tokens[0].c_str(), 0, 0);
+  unsigned int event = strtoul(tokens[1].c_str(), 0, 0);
   to = edm::EventID(run, event);
   return true;
 }  // decode to EventID
